backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
aiogram==3.16.0
aiohttp==3.11.11
groq==0.13.1
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.4
